# Conflicts:
#	.appveyor.yml
#	.bash_helpers.sh
#	.travis.yml
#	CMake/Modules/CheckCompilerToolchain.cmake
#	CMake/Modules/FindDirectX.cmake
#	CMake/Modules/FindUrho3D.cmake
#	CMake/Modules/UrhoCommon.cmake
#	CMakeLists.txt
#	Docs/AngelScriptAPI.h
#	Docs/Doxyfile.in
#	Docs/GettingStarted.dox
#	Docs/LuaScriptAPI.dox
#	Docs/Reference.dox
#	Docs/ScriptAPI.dox
#	Docs/Urho3D.dox
#	README.md
#	Rakefile
#	Resources/EditorData/AtomicEditor/eulas/atomic_thirdparty_eula.txt
#	Source/Atomic/Atomic2D/ParticleEmitter2D.cpp
#	Source/Atomic/Atomic2D/PhysicsWorld2D.cpp
#	Source/Atomic/Atomic2D/SpriteSheet2D.cpp
#	Source/Atomic/Atomic2D/TmxFile2D.cpp
#	Source/Atomic/Audio/Sound.h
#	Source/Atomic/Core/Context.cpp
#	Source/Atomic/Core/Main.h
#	Source/Atomic/Core/ProcessUtils.cpp
#	Source/Atomic/Core/ProcessUtils.h
#	Source/Atomic/Engine/Application.cpp
#	Source/Atomic/Engine/Application.h
#	Source/Atomic/Engine/Engine.cpp
#	Source/Atomic/Graphics/Animation.h
#	Source/Atomic/Graphics/Model.cpp
#	Source/Atomic/Graphics/Model.h
#	Source/Atomic/Graphics/Text3D/Text3D.cpp
#	Source/Atomic/Graphics/Text3D/Text3D.h
#	Source/Atomic/Graphics/Text3D/Text3DBatch.cpp
#	Source/Atomic/Graphics/Text3D/Text3DBatch.h
#	Source/Atomic/Graphics/Text3D/Text3DBitmap.cpp
#	Source/Atomic/Graphics/Text3D/Text3DFont.cpp
#	Source/Atomic/Graphics/Text3D/Text3DFont.h
#	Source/Atomic/Graphics/Text3D/Text3DFontFace.cpp
#	Source/Atomic/Graphics/Text3D/Text3DFreeType.cpp
#	Source/Atomic/Graphics/Text3D/Text3DText.cpp
#	Source/Atomic/Graphics/Text3D/Text3DText.h
#	Source/Atomic/Graphics/Texture.h
#	Source/Atomic/Graphics/VertexBuffer.cpp
#	Source/Atomic/IO/Log.cpp
#	Source/Atomic/Physics/PhysicsWorld.cpp
#	Source/Atomic/Resource/JSONFile.cpp
#	Source/Atomic/Scene/Node.h
#	Source/CMakeLists.txt
#	Source/Samples/40_Localization/L10n.cpp
#	Source/Samples/42_PBRMaterials/PBRMaterials.cpp
#	Source/Samples/45_InverseKinematics/CMakeLists.txt
#	Source/ThirdParty/Civetweb/CMakeLists.txt
#	Source/ThirdParty/Lua/CMakeLists.txt
#	Source/ThirdParty/Lua/src/loslib.c
#	Source/ThirdParty/LuaJIT/CMakeLists.txt
#	Source/ThirdParty/LuaJIT/src/lj_arch.h
#	Source/ThirdParty/SQLite/CMakeLists.txt
#	Source/ThirdParty/boost/preprocessor/cat.hpp
#	Source/ThirdParty/boost/preprocessor/config/config.hpp
#	Source/ThirdParty/boost/preprocessor/seq/detail/binary_transform.hpp
#	Source/ThirdParty/boost/preprocessor/seq/to_list.hpp
#	Source/ThirdParty/rapidjson/include/rapidjson/internal/dtoa.h
#	Source/ThirdParty/rapidjson/include/rapidjson/internal/ieee754.h
#	Source/ThirdParty/rapidjson/include/rapidjson/internal/regex.h
#	Source/ThirdParty/rapidjson/include/rapidjson/internal/strtod.h
#	Source/ThirdParty/rapidjson/include/rapidjson/istreamwrapper.h
#	Source/ThirdParty/rapidjson/include/rapidjson/pointer.h
#	Source/ThirdParty/rapidjson/include/rapidjson/schema.h
#	Source/Tools/AssetImporter/AssetImporter.cpp
#	Source/Tools/CMakeLists.txt
#	Source/Tools/PackageTool/PackageTool.cpp
#	Source/Tools/Urho3DPlayer/CMakeLists.txt
#	Source/Tools/Urho3DPlayer/Urho3DPlayer.cpp
#	Source/Urho3D/.soversion
#	Source/Urho3D/AngelScript/APITemplates.h
#	Source/Urho3D/AngelScript/AudioAPI.cpp
#	Source/Urho3D/AngelScript/CoreAPI.cpp
#	Source/Urho3D/AngelScript/GraphicsAPI.cpp
#	Source/Urho3D/AngelScript/IKAPI.cpp
#	Source/Urho3D/AngelScript/IOAPI.cpp
#	Source/Urho3D/AngelScript/InputAPI.cpp
#	Source/Urho3D/AngelScript/MathAPI.cpp
#	Source/Urho3D/AngelScript/PhysicsAPI.cpp
#	Source/Urho3D/AngelScript/ResourceAPI.cpp
#	Source/Urho3D/AngelScript/ScriptAPI.h
#	Source/Urho3D/AngelScript/ScriptAPIDump.cpp
#	Source/Urho3D/AngelScript/ScriptFile.cpp
#	Source/Urho3D/AngelScript/UIAPI.cpp
#	Source/Urho3D/AngelScript/Urho2DAPI.cpp
#	Source/Urho3D/CMakeLists.txt
#	Source/Urho3D/LuaScript/pkgs/Audio/Sound.pkg
#	Source/Urho3D/LuaScript/pkgs/Audio/SoundSource.pkg
#	Source/Urho3D/LuaScript/pkgs/Core/ProcessUtils.pkg
#	Source/Urho3D/LuaScript/pkgs/Graphics/Animation.pkg
#	Source/Urho3D/LuaScript/pkgs/Graphics/Model.pkg
#	Source/Urho3D/LuaScript/pkgs/Graphics/Texture.pkg
#	Source/Urho3D/LuaScript/pkgs/Input/Input.pkg
#	Source/Urho3D/LuaScript/pkgs/Math/Quaternion.pkg
#	Source/Urho3D/LuaScript/pkgs/Math/Vector3.pkg
#	Source/Urho3D/LuaScript/pkgs/PhysicsLuaAPI.pkg
#	Source/Urho3D/LuaScript/pkgs/Resource/JSONValue.pkg
#	Source/Urho3D/LuaScript/pkgs/Resource/Resource.pkg
#	Source/Urho3D/LuaScript/pkgs/Scene/Node.pkg
#	Source/Urho3D/LuaScript/pkgs/UI/Font.pkg
#	Source/Urho3D/LuaScript/pkgs/UI/Text.pkg
#	Source/Urho3D/LuaScript/pkgs/UI/Text3D.pkg
#	Source/Urho3D/LuaScript/pkgs/UI/UI.pkg
#	Source/Urho3D/LuaScript/pkgs/Urho2D/ParticleEmitter2D.pkg
#	Source/Urho3D/UI/Cursor.cpp
#	Source/Urho3D/UI/UI.cpp
#	Source/Urho3D/UI/UI.h
#	Source/Urho3D/Urho2D/PhysicsEvents2D.h
#	bin/Autoload/LargeData/Materials/PBR/Check.xml
#	bin/Autoload/LargeData/Materials/PBR/HighRoughMetallic0.xml
#	bin/Autoload/LargeData/Materials/PBR/HighRoughMetallic10.xml
#	bin/Autoload/LargeData/Materials/PBR/HighRoughMetallic3.xml
#	bin/Autoload/LargeData/Materials/PBR/HighRoughMetallic5.xml
#	bin/Autoload/LargeData/Materials/PBR/HighRoughMetallic7.xml
#	bin/Autoload/LargeData/Materials/PBR/Metallic0.xml
#	bin/Autoload/LargeData/Materials/PBR/Metallic10.xml
#	bin/Autoload/LargeData/Materials/PBR/Metallic3.xml
#	bin/Autoload/LargeData/Materials/PBR/Metallic5.xml
#	bin/Autoload/LargeData/Materials/PBR/Metallic7.xml
#	bin/Autoload/LargeData/Materials/PBR/MetallicRough0.xml
#	bin/Autoload/LargeData/Materials/PBR/MetallicRough10.xml
#	bin/Autoload/LargeData/Materials/PBR/MetallicRough3.xml
#	bin/Autoload/LargeData/Materials/PBR/MetallicRough5.xml
#	bin/Autoload/LargeData/Materials/PBR/MetallicRough7.xml
#	bin/Autoload/LargeData/Materials/PBR/Mud.xml
#	bin/Autoload/LargeData/Materials/PBR/Roughness0.xml
#	bin/Autoload/LargeData/Materials/PBR/Roughness10.xml
#	bin/Autoload/LargeData/Materials/PBR/Roughness3.xml
#	bin/Autoload/LargeData/Materials/PBR/Roughness5.xml
#	bin/Autoload/LargeData/Materials/PBR/Roughness7.xml
#	bin/Autoload/LargeData/Materials/PBR/Tile.xml
#	bin/Data/LuaScripts/40_Localization.lua
#	bin/Data/LuaScripts/42_PBRMaterials.lua
#	bin/Data/Scenes/PBRExample.xml
#	bin/Data/Scripts/40_Localization.as
#	bin/Data/Scripts/42_PBRMaterials.as
#	bin/Data/Scripts/Editor.as
#	bin/Data/Scripts/Editor/EditorUI.as
#	bin/Data/Scripts/Editor/EditorView.as
#	cmake_generic.sh

+++ Source/Atomic/Graphics/Text3D/Text3DBitmap.h
@@ -42,7 +42,7 @@ public:
     ~Text3DBitmap();
 
     /// Load font face.
-    virtual bool Load(const unsigned char* fontData, unsigned fontDataSize, int pointSize);
+    virtual bool Load(const unsigned char* fontData, unsigned fontDataSize, float pointSize);
     /// Load from existed font face, pack used glyphs into smallest texture size and smallest number of texture.
     bool Load(Text3DFontFace* fontFace, bool usedGlyphs);
     /// Save as a new bitmap font type in XML format. Return true if successful.
