+++ Source/Atomic/Graphics/Text3D/Text3DFontFace.cpp
@@ -70,7 +70,7 @@ const Text3DFontGlyph* Text3DFontFace::GetGlyph(unsigned c)
         return 0;
 }
 
-short Text3DFontFace::GetKerning(unsigned c, unsigned d) const
+float Text3DFontFace::GetKerning(unsigned c, unsigned d) const
 {
     if (kerningMapping_.Empty())
         return 0;
@@ -83,7 +83,7 @@ short Text3DFontFace::GetKerning(unsigned c, unsigned d) const
 
     unsigned value = (c << 16) + d;
 
-    HashMap<unsigned, short>::ConstIterator i = kerningMapping_.Find(value);
+    HashMap<unsigned, float>::ConstIterator i = kerningMapping_.Find(value);
     if (i != kerningMapping_.End())
         return i->second_;
